requirements.txt
python-dotenv==1.0.0
langchain==0.0.191
pinecone-client==2.2.1
black==23.3.0
urllib3==1.26.6
unstructured==0.7.1
tabulate==0.9.0
pdf2image==1.16.3
openai==0.27.7